requirements.txt
asgiref==3.4.1
Django==3.2.11
django-tastypie==0.14.4
django-tastypie-extras==0.4b4
et-xmlfile==1.1.0
gunicorn==20.1.0
openpyxl==3.0.9
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
whitenoise==5.3.0
